refactor(maestro): split ide/completion/template into cohesive modules
